Add register form layout
