+++ src/Components/Register/registerForm.css
@@ -1,0 +1,4 @@
+.label-font{
+    font-size: 12px;
+    color: gray.500;
+}
